feat(find_court): Fine tune courts db
Add partial match boolean and default to false

Courts-DB was initially given a broader ability to
parse thru and find partial matches when nothing
actually hit.

As a result the tool was overly excited to find
partial matches .. at the very least.  This is useful
if you give a long sentence
and it needs to find one particular court
but this should not be the default